requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.1.3
packaging==24.2
pandas==2.2.3
plotly==5.24.1
PyMySQL==1.1.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tenacity==9.0.0
tzdata==2024.2
Werkzeug==3.1.3
